Add support for setting UID/GID through mount options
Even with read-only support, it may be useful to set the filesystem
default UID/GID to an arbitrary value.

With this patch, the user can use the "uid=N", "gid=N", "user_id=N" or
"group_id=N" options to set the user and group IDs for the whole
filesystem.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>